Address review feedback
